rename to electronStore.ts
