+++ src/background/electronStore.ts
